requirements.txt
appnope==0.1.3
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
idna==3.4
imageio==2.32.0
imutils==0.5.4
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.4
matplotlib==3.8.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.2.1
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
pyzmq==25.0.0
requests==2.28.2
scikit-image==0.21.0
scipy==1.11.3
setuptools==65.6.3
six==1.16.0
tifffile==2023.9.26
torch==2.1.1
torchvision==0.16.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
wget==3.2
wheel==0.37.1